wip! Rework sha3_update_aligned into multiple functions
Add a stupid, over-engineered, probably-broken, unfriendly to 32-bit
architectures, rate-based update function for when the stars align and
we have a rate-sized chunk of suitably-aligned bytes to read in.

Rewrite the old sha3_update_aligned() to take pointers to the input
buffer and the input length to let it return to sha3_update so we can
clean up any remaining bytes via the slow method.