Replace sha3_update_aligned()
Rather than calling another function to update the SHA-3 state with
aligned data, do it inline. This saves a function call and - more
importantly - allows the aligned code path to be used at the start of
input of arbitrary length.

The sha3_update() function can now update the state with a rate-sized
block at a time - depending on input alignment and position in the state
array - and 8 bytes at a time if both the input and position are
aligned.

This has only been tested on 64-bit architectures and is probably
unfriendly to 32-bit. It may be worth separating the body of the various
loops on those systems.